Changed ints to double Part I

+++ src/com/hexcore/cas/model/Cell.java
@@ -2,28 +2,28 @@ package com.hexcore.cas.model;
 
 public class Cell
 {
-	private int[] 	values;
+	private double[] 	values;
 	
 	public Cell(int valueCount)
 	{
-		values = new int[valueCount];
+		values = new double[valueCount];
 		
-		for (int i = 0; i < valueCount; i++) values[i] = 0;
+		for (int i = 0; i < valueCount; i++) values[i] = 0.0;
 	}
 	
-	public Cell(int[] values)
+	public Cell(double[] values)
 	{
-		this.values = (int[])values.clone();
+		this.values = (double[])values.clone();
 	}
 	
 	public Cell(Cell cell)
 	{
-		this.values = (int[])cell.values.clone();
+		this.values = (double[])cell.values.clone();
 	}
 	
-	public int getValue(int index)
+	public double getValue(int index)
 	{
-		if ((index < 0) || (index >= values.length)) return 0;
+		if ((index < 0) || (index >= values.length)) return 0.0;
 		return values[index];
 	}
 	
@@ -32,12 +32,12 @@ public class Cell
 		return values.length;
 	}
 	
-	public int[] getValues()
+	public double[] getValues()
 	{
 		return values;
 	}
 	
-	public void setValue(int index, int value)
+	public void setValue(int index, double value)
 	{
 		if ((index < 0) || (index >= values.length)) return;
 		values[index] = value;
