+++ src/com/hexcore/cas/model/test/TestHexagonGrid.java
@@ -50,7 +50,7 @@ public class TestHexagonGrid extends TestCase
 			for(int j = 0; j < x; j++)
 			{
 				Cell[] neighbours = grid.getNeighbours(new Vector2i(j,i));
-				int temp = 0;
+				double temp = 0.0;
 				for(int k = 0; k < 6; k++)
 				{
 					if(neighbours[k] != null)
